Backend/requirements.txt
# ============================
# Production dependencies
# ============================
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.3.6
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.4
distro==1.9.0
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-extensions==4.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kombu==5.3.4
MarkupSafe==3.0.2
openai==1.93.0
packaging==25.0
pdfminer.six==20250506
Pillow==11.3.0
prompt-toolkit==3.0.51
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
PyMuPDF==1.26.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
redis==4.5.5
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13

# ============================
# Development / Tooling dependencies
# ============================
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
ipython==8.12.3
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pip-check==3.2.1
pipdeptree==2.28.0
pipreqs==0.5.0
platformdirs==4.4.0
ptyprocess==0.7.0
pure-eval==0.2.3
Pygments==2.19.2
pyzmq==27.0.2
referencing==0.36.2
rpds-py==0.27.1
setuptools==59.6.0
soupsieve==2.8
stack-data==0.6.3
terminaltables==3.1.10
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
webencodings==0.5.1
yarg==0.1.9
